% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GeneOntology.R
\name{aggrgate_GO_offspring}
\alias{aggrgate_GO_offspring}
\title{Aggregate GO offspring genesets and genes}
\usage{
aggrgate_GO_offspring(gs.GO)
}
\arguments{
\item{gs.GO}{GeneSetCollection from getGenesets()}
}
\value{
GeneSetCollection with aggregated gene sets
}
\description{
Follow the GO hierarchy down and include all genes in offspring sets for a given gene set
}
