% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corInGeneSet.R
\name{corInGeneSet}
\alias{corInGeneSet}
\title{Evaluate mean correlation between residuals in gene set}
\usage{
corInGeneSet(fit, idx, squareCorr = FALSE)
}
\arguments{
\item{fit}{result of differential expression with dream}

\item{idx}{indeces or rownames to extract}

\item{squareCorr}{compute the mean squared correlation instead}
}
\description{
Evaluate mean correlation between residuals in gene set based on results from dream
}
