% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith.R
\name{.rankSumTestWithCorrelation}
\alias{.rankSumTestWithCorrelation}
\title{Two Sample Wilcoxon-Mann-Whitney Rank Sum Test Allowing For Correlation}
\usage{
.rankSumTestWithCorrelation(index, statistics, correlation = 0, df = Inf)
}
\arguments{
\item{index}{any index vector such that 'statistics[index]' contains the values of the statistic for the test group.}

\item{statistics}{numeric vector giving values of the test statistic.}

\item{correlation}{numeric scalar, average correlation between cases in the test group.  Cases in the second group are assumed independent of each other and other the first group.}

\item{df}{degrees of freedom which the correlation has been estimated.}
}
\description{
Same as \code{limma::.rankSumTestWithCorrelation}, but returns effect size.
}
\details{
See \code{limma::.rankSumTestWithCorrelation}
}
