% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeToList.R
\name{recodeToList,GeneSetCollection-method}
\alias{recodeToList,GeneSetCollection-method}
\title{Recode GeneSetCollection to list used by limma}
\usage{
\S4method{recodeToList}{GeneSetCollection}(gsc)
}
\arguments{
\item{gsc}{GeneSetCollection object}
}
\description{
Recode GeneSetCollection to list used by limma
}
