% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenith.R
\name{zenith}
\alias{zenith}
\title{Gene set analysis following differential expression with dream}
\usage{
zenith(
  fit,
  coef,
  index,
  use.ranks = FALSE,
  allow.neg.cor = FALSE,
  squaredStats = FALSE
)
}
\arguments{
\item{fit}{result of differential expression with dream}

\item{coef}{coefficient to test using topTable(fit, coef)}

\item{index}{an index vector or a list of index vectors.  Can be any vector such that 'fit[index,]'  selects the rows corresponding to the test set.  The list can be made using 'ids2indices'.}

\item{use.ranks}{do a rank-based test ('TRUE') or a parametric test ('FALSE')?}

\item{allow.neg.cor}{should reduced variance inflation factors be allowed for negative correlations?}

\item{squaredStats}{Test squared test statstics to identify gene sets with log fold change of mixed sign.}
}
\description{
Perform gene set analysis on the result of differential expression using linear (mixed) modeling with \code{variancePartition::dream} by considering the correlation between gene expression traits.  This package is a slight modification of \code{limma::camera} to 1) be compatible with dream, and 2) allow identification of gene sets with log fold changes with mixed sign.
}
\details{
For differential expression with dream using linear (mixed) models see Hoffman and Roussos (2020).  For the original camera gene set test see Wu and Smyth (2012).
}
\references{
{
  \insertRef{hoffman2020dream}{zenith}

  \insertRef{wu2012camera}{zenith}
}
}
