% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zenithPR_gsa.R
\name{zenithPR_gsa}
\alias{zenithPR_gsa}
\title{Gene set analysis using pre-computed test statistic}
\usage{
zenithPR_gsa(
  statistics,
  ids,
  geneSets,
  use.ranks = FALSE,
  n_genes_min = 10,
  progressbar = TRUE,
  inter.gene.cor = 0.01
)
}
\arguments{
\item{statistics}{pre-computed test statistics}

\item{ids}{name of gene for each entry in \code{statistics}}

\item{geneSets}{\code{GeneSetCollection}}

\item{use.ranks}{do a rank-based test \code{TRUE} or a parametric test \code{FALSE}? default: FALSE}

\item{n_genes_min}{minumum number of genes in a geneset}

\item{progressbar}{if TRUE, show progress bar}

\item{inter.gene.cor}{correlation of test statistics with in gene set}
}
\value{
\itemize{
  \item \code{NGenes}: number of genes in this set
  \item \code{Correlation}: mean correlation between expression of genes in this set
  \item \code{delta}: difference in mean t-statistic for genes in this set compared to genes not in this set
  \item \code{se}: standard error of \code{delta}
  \item \code{p.less}: p-value for hypothesis test of \code{H0: delta < 0}
  \item \code{p.greater}: p-value for hypothesis test of \code{H0: delta > 0}
  \item \code{PValue}:  p-value for hypothesis test \code{H0: delta != 0}
  \item \code{Direction}: direction of effect based on sign(delta)
  \item \code{FDR}: false discovery rate based on Benjamini-Hochberg method in \code{p.adjust}
}
}
\description{
Perform gene set analysis on the result of a pre-computed test statistic.  Test whether statistics in a gene set are larger/smaller than statistics not in the set.
}
\details{
This is the same as \code{zenith_gsa()}, but uses pre-computed test statistics.  Note that \code{zenithPR_gsa()} may give slightly different results for small samples sizes, if \code{zenithPR_gsa()} is fed t-statistics instead of z-statistics.
}
\seealso{
\code{zenith_gsa()}
}
