% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MSigDB.R
\name{get_MSigDB}
\alias{get_MSigDB}
\title{Load MSigDB genesets}
\usage{
get_MSigDB(
  cat = unique(msigdbr_collections()$gs_cat),
  to = "ENSEMBL",
  org = "hsa"
)
}
\arguments{
\item{cat}{array of categories to load.  Defaults to array of all MSigDB categories}

\item{to}{convert gene names to this type using \code{EnrichmentBrowser::idMap()}.  See \code{EnrichmentBrowser::idTypes(org="hsa")} for valid types}

\item{org}{organism.  human (\code{'hsa'}), mouse (\code{'mmu'}), etc}
}
\value{
Gene sets stored as GeneSetCollection
}
\description{
Load MSigDB genesets
}
\details{
This function loads the MSigDB gene sets using the packages \code{EnrichmentBrowser} and \code{msigdbr}.  It can take a mintute to load because converting gene name type is slow.
}
\examples{
# load Hallmark gene sets
gs = get_MSigDB('H')

# load all gene sets
# gs = get_MSigDB()

}
