% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MSigDB.R
\name{get_MSigDB}
\alias{get_MSigDB}
\title{Load MSigDB genesets}
\usage{
get_MSigDB(
  cat = c("H", "C1", "C2", "C3", "C4", "C5", "C6", "C7"),
  to = "ENSEMBL"
)
}
\arguments{
\item{cat}{array of categories to load}

\item{to}{convert gene names to this type using \code{EnrichmentBrowser::idMap()}.  See \code{EnrichmentBrowser::idTypes(org="hsa")} for valid types}
}
\value{
Gene sets stored as GeneSetCollection
}
\description{
Load MSigDB genesets
}
\details{
This function loads the MSigDB gene sets using the packages \code{EnrichmentBrowser} and \code{msigdbr}.  It can take a mintute to load because converting gene name type is slow.
}
\examples{
# load Hallmark gene sets
gs = get_MSigDB('H')

# load all gene sets
# gs = get_MSigDB()

}
