% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeToList.R
\name{recodeToList}
\alias{recodeToList}
\title{Recode GeneSetCollection to list used by limma}
\usage{
recodeToList(gsc)
}
\arguments{
\item{gsc}{GeneSetCollection object}
}
\value{
list storing data from GeneSetCollection
}
\description{
Recode \code{GeneSetCollection} to \code{list} used by \code{limma}
}
\examples{

# load Hallmark gene sets
gs = get_MSigDB('H')

# recode GeneSetCollection as a list
gs.list = recodeToList(gs)

}
