% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recodeToList.R
\name{recodeToList}
\alias{recodeToList}
\title{Recode GeneSetCollection to list used by limma}
\usage{
recodeToList(gsc)
}
\arguments{
\item{gsc}{GeneSetCollection object}
}
\description{
Recode \code{GeneSetCollection} to \code{list} used by \code{limma}
}
